21: Updated MergedMatchInfo to a sql view
